requirements.txt
psycopg2-binary
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic
